Fix various Linux issues
Move 'run' code to env; make get_file_version work on Linux

Moves run_exe to env wrappers and makes it return a Popen reference
we can access from within app_buttons or wherever else a subprocess is called.

Reimplements get_file_version to use native binary reading utilities on Linux. It Just Works!

Refines Java detection a bit.

Added some type hints and comments.

Infernio:

Some minor followup fixes for Linux

- Prefer JAVA_HOME if set
- Skip 0.0.0.0 and 1.0.0.0 versions to match the Windows version
- Fix reading of xSE versions (uses commas and spaces instead of dots)
- Get rid of u'' prefixes and use f-strings

Deprint about missing 'strings' program and add a cache

It's not *that* slow (adds up to about 0.02s on boot on my Linux
system), but there's no reason to keep reading the same files over and
over just in case the version changed. Added a cache decorator to the
Windows version too, since the same consideration applies there, the
pywin32 calls are cheap, but not *that* cheap.

Utumno:

ExeLauncher is used for .exe files with the exception of LOOT - best thing
here is to forget about special handling of LOOT as one can always add
the unix path to bash.ini. So I removed this particularly nasty override
- whenever you see an os. check outside env is nasty - and I moved the
_run_exe code for non .exe files to linux.AppLauncher.launch_app.
Untested but should work

- move allow_create to base and disallow handling of .lnk files in linux.
This is the base implementation, we can drop return False now that linux
is fleshed out a bit :)

xdg-open reported not to take any args with or without the '--' - nit in
get_file_version -> avoid looping twice, hopefully 'FileVersion' and co
come early.

Under #243

+++ Mopy/bash/env/common.py
@@ -128,6 +128,14 @@ def _parse_vdf(vdf_path: _StrPath, *, vdf_root: str) -> dict | None:
         return None
     return vdf_root
 
+def _parse_version_string(full_ver):
+    # xSE uses commas in its version fields, so use this 'heuristic'
+    split_on = ',' if ',' in full_ver else '.'
+    try:
+        return tuple([int(part) for part in full_ver.split(split_on)])
+    except ValueError:
+        return 0, 0, 0, 0
+
 @functools.cache
 def _get_steam_manifests(steam_path: _StrPath) -> dict[int, str]:
     """Read the libraryfolders.vdf file used by Steam for storing the Steam
@@ -435,7 +443,9 @@ def file_operation(operation: FileOperationType,
                     deprint(f'Permission to create trash directory denied, '
                             f'permanently deleting the file ({to_delete}) '
                             f'instead', traceback=True)
-            if to_delete.is_dir():
+            if to_delete.is_dir() and not os.path.islink(to_delete):
+                # python would attempt to call rmtree on symlinks to dirs, and
+                # raise
                 to_delete.rmtree(to_delete.stail)
             else:
                 to_delete.remove()
@@ -467,7 +477,7 @@ def set_cwd(func):
         cwd = os.getcwd()
         os.chdir(exe_path.head.s)
         try:
-            func(self, exe_path, *args)
+            return func(self, exe_path, *args)
         finally:
             os.chdir(cwd)
     return _switch_dir
@@ -485,7 +495,8 @@ class _AppLauncher:
         self._display_launcher = display_launcher
         self._exe_args = cli_args
 
-    def allow_create(self): return False ##: for linux - flesh out!
+    def allow_create(self): # if self.exePath does not exist this must be False
+        return self._display_launcher
 
     @classmethod
     def find_launcher(cls, app_exe, app_key, *, root_dirs: tuple | str = tuple(
